Chad-DevOps/nessus/requirements.txt
arrow==1.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
idna==3.3
marshmallow==3.17.0
packaging==21.3
pendulum==2.1.2
pyparsing==3.0.9
pyTenable==1.4.7
python-box==6.0.2
python-dateutil==2.8.2
pytzdata==2020.1
requests==2.28.1
restfly==1.4.6
semver==2.13.0
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
